requirements.txt
fastapi
uvicorn
PyPDF2
passlib
openai
pinecone-client
python-dotenv
httpx
